Create custom hook 'useTasks' to retrieve tasks.
