+++ frontend/src/axios.js
@@ -1,5 +1,5 @@
 import axios from "axios";
 
 export const axiosConn = axios.create({
-  baseURL: "http://localhost:8000/"
+  baseURL: "http://localhost:8000/api/"
 });
